[Fix] #37 - WITextField가 목표일수 입력 경우에서도 활용될 수 있게 코드 수정

+++ Winey/Winey/Source/Scenes/Upload/Model/UploadModel.swift
@@ -10,5 +10,5 @@ import UIKit
 struct UploadModel: Hashable {
     let feedImage: UIImage
     let feedTitle: String
-    let feedPrcie: Int64
+    let feedPrcie: Int
 }
